lock all page for watchdogd
BZ 37868

when the system is in a low memory state, clean pages are freed
from the page cache to allow processes to allocate memory even
if no swap is available. The goal of the mlockall function is
to deactivate the paging for a process, the swapping deactivation
for this process is to be considered as a side effect.
Currently, in our system, it will trigger a warning WDT when
watchdogd demon didn¡¯t feed watchdog in 10 seconds. This
warning wdt will trigger a panic.But there is a ugly case, which
will cause watchdog demon won¡¯t feed the WDT. When system is
hungry of memory, it will reclaim the clean page of watchdog demon,
in such case, watchdog demon has no chance to feed the WDT. we need
to let the system has time to reclaim the memory. So clean page of
watchdog should not reclaim. We meet many case like this,
that will trigger WDT.
Signed-off-by: xiaobing tu <xiaobing.tu@intel.com>

Change-Id: I6c9843b319800d4961e31d07f76e860ac74aafc9
Signed-off-by: xiaobing tu <xiaobing.tu@intel.com>
Reviewed-on: http://android.intel.com:8080/50314
Reviewed-by: Gross, Mark <mark.gross@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Reviewed-by: Allyn, Mark A <mark.a.allyn@intel.com>
Tested-by: Ng, Cheon-woei <cheon-woei.ng@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>